Handle negstive

+++ libs/server/GlobUtils.cs
@@ -16,7 +16,7 @@ namespace Garnet.server
         {
             static byte ToLowerAscii(byte value)
             {
-                if ((value - 'A') <= ('Z' - 'A')) // Is in [A-Z]
+                if ((uint)(value - 'A') <= (uint)('Z' - 'A')) // Is in [A-Z]
                     value = (byte)(value | 0x20);
                 return value;
             }
